Fixes to docs
