+++ src/io/ipc/write/file_async.rs
@@ -15,7 +15,7 @@ use crate::io::ipc::{IpcField, ARROW_MAGIC};
 
 type WriteOutput<W> = (usize, Option<Block>, Vec<Block>, Option<W>);
 
-///  Sink that writes array [`chunks`](Chunk) as an IPC file.
+///  Sink that writes array [`chunks`](crate::chunk::Chunk) as an IPC file.
 ///
 /// The file header is automatically written before writing the first chunk, and the file footer is
 /// automatically written when the sink is closed.
